requirements.txt
# Audio capture
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.24.0

# Local transcription (Apple Silicon Metal via CTranslate2)
faster-whisper>=1.0.0

# Local LLM inference
ollama>=0.3.0

# CLI framework
typer[all]>=0.12.0
rich>=13.7.0

# Word document generation
python-docx>=1.1.0

# Data validation & config
pydantic>=2.5.0
pyyaml>=6.0.1

# Retry logic for Ollama calls
tenacity>=8.2.0
